requirements.txt
# Core ML/Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Cheminformatics
rdkit>=2023.9.1

# Web App
streamlit>=1.28.0

# Image processing
pillow>=10.0.0
matplotlib>=3.7.0

# API requests
requests>=2.31.0

# Data download
urllib3>=2.0.0
